Classes Adjustments
Deleted BestRouteWrapper then changed best_route_wrapper to  BestRouteWrapper. Adjustments in Main_BestRoute.
